ref: Move TextToAmmount from parsing to userinput
