+++ internal/userinput/userinput.go
@@ -4,6 +4,7 @@ package userinput
 import (
 	"fmt"
 
+	"github.com/shopspring/decimal"
 	"github.com/vitorqb/addledger/internal/finance"
 	"github.com/vitorqb/addledger/internal/journal"
 	"github.com/vitorqb/addledger/internal/state"
@@ -182,3 +183,22 @@ const (
 	Context DoneSource = "context"
 	Input   DoneSource = "input"
 )
+
+func TextToAmmount(x string) (finance.Ammount, error) {
+	var err error
+	var quantity decimal.Decimal
+	var commodity string
+	switch words := strings.Split(x, " "); len(words) {
+	case 1:
+		quantity, err = decimal.NewFromString(words[0])
+	case 2:
+		commodity = words[0]
+		quantity, err = decimal.NewFromString(words[1])
+	default:
+		return finance.Ammount{}, fmt.Errorf("invalid format")
+	}
+	if err != nil {
+		return finance.Ammount{}, fmt.Errorf("invalid format: %w", err)
+	}
+	return finance.Ammount{Commodity: commodity, Quantity: quantity}, nil
+}
