+++ internal/userinput/userinput_test.go
@@ -5,7 +5,10 @@ import (
 	"testing"
 
 	"fmt"
+
+	"github.com/shopspring/decimal"
 	"github.com/stretchr/testify/assert"
+	"github.com/vitorqb/addledger/internal/finance"
 	"github.com/vitorqb/addledger/internal/journal"
 	"github.com/vitorqb/addledger/internal/state"
 	"github.com/vitorqb/addledger/internal/testutils"
@@ -337,5 +340,75 @@ func TestTagTextToTag__Bad(t *testing.T) {
 			t.Errorf(fmt.Sprintf("Expected error, got none: %s", input))
 		}
 
+	}
+}
+
+func TestTextToAmmount(t *testing.T) {
+	type testcase struct {
+		text     string
+		ammount  finance.Ammount
+		errorMsg string
+	}
+	var testcases = []testcase{
+		{
+			text: "EUR 12.20",
+			ammount: finance.Ammount{
+				Commodity: "EUR",
+				Quantity:  decimal.New(1220, -2),
+			},
+		},
+		{
+			text: "EUR 99999.99999",
+			ammount: finance.Ammount{
+				Commodity: "EUR",
+				Quantity:  decimal.NewFromFloat(99999.99999),
+			},
+		},
+		{
+			text: "12.20",
+			ammount: finance.Ammount{
+				Commodity: "",
+				Quantity:  decimal.New(1220, -2),
+			},
+		},
+		{
+			text:     "12,20",
+			errorMsg: "invalid format",
+		},
+		{
+			text:     "EUR",
+			errorMsg: "invalid format",
+		},
+		{
+			text:     "EUR 12 12",
+			errorMsg: "invalid format",
+		},
+		{
+			text:     "12 FOO",
+			errorMsg: "invalid format",
+		},
+		{
+			text:     "EUR  12.20",
+			errorMsg: "invalid format",
+		},
+		{
+			text:     "EUR 12.20 ",
+			errorMsg: "invalid format",
+		},
+		{
+			text:     " EUR 12.20 ",
+			errorMsg: "invalid format",
+		},
+	}
+	for _, tc := range testcases {
+		t.Run(tc.text, func(t *testing.T) {
+			result, err := TextToAmmount(tc.text)
+			if tc.errorMsg == "" {
+				assert.Nil(t, err)
+				assert.Equal(t, tc.ammount, result)
+			} else {
+				assert.ErrorContains(t, err, tc.errorMsg)
+			}
+		})
 	}
 }
